requirements.txt
asgiref==3.2.10
attrs==19.3.0
cycler==0.10.0
decorator==4.4.2
Django==3.0.8
django-model-utils==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.2
more-itertools==8.4.0
networkx==2.4
numpy==1.19.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.0
six==1.15.0
sqlparse==0.3.1
tifffile==2020.6.3
wcwidth==0.2.5
